requirements_mac.txt
asttokens==2.4.1
black==24.8.0
click==8.1.7
decorator==5.1.1
diskcache==5.6.3
django-environ==0.9.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.15.4
fsspec==2024.6.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==2.0.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plyfile==1.0.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tomli==2.0.1
torch==2.5.0.dev20240902
torchaudio==2.5.0.dev20240902
torchvision==0.20.0.dev20240902
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
